// src/components/DetailsCard/interface.ts
interface Genre {
  id: number;
  name: string;
}

export default interface Movie {
  id: number;
  title: string;
  overview: string;
  release_date: string | null;
  vote_average: number;
  parsedVoteAverage: string;
  vote_count: number;
  poster_path: string | null;
  image: string | null;
  genres: Genre[];
  runtime: number | null;
}